+++ account-service/src/main/java/com/bilgeadam/banking/account_service/domain/AccountType.java
@@ -2,5 +2,5 @@ package com.bilgeadam.banking.account_service.domain;
 
 public enum AccountType {
     SAVINGS,
-    CHECKİNG
+    CHECKING
 }
